Add mongoose chart repository.
