+++ console/src/app/core/models/chart/mongoose-chart-interface/mongoose-chart.interface.ts
@@ -12,4 +12,5 @@ export interface MongooseChart {
 
     mongooseChartDao: MongooseChartDao; 
     updateChart(recordLoadStepId: string); 
+    shouldDrawChart(): boolean; 
 }
